requirements/requirements.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0

# Django REST Framework
djangorestframework==3.12.2

# Passwords security
argon2-cffi==20.1.0

# Debugging
ipdb==0.11

# Tools
django-extensions==3.0.9

# Code quality
flake8==3.6.0